Added methods for both android, ios retrieveSubscribed(), retrievePlayerId(), retrieveToken().
