+++ android/manifest
@@ -4,7 +4,7 @@
 #
  #
 
-version: 2.1.1
+version: 2.1.2
 apiversion: 4
 architectures: arm64-v8a armeabi-v7a x86
 description: com.williamrijksen.onesignal
@@ -17,4 +17,4 @@ name: com.williamrijksen.onesignal
 moduleid: com.williamrijksen.onesignal
 guid: 67065763-fd5e-4069-a877-6c7fd328f877
 platform: android
-minsdk: 7.5.0.GA
+minsdk: 8.3.1.GA
