+++ ios/Cartfile.resolved
@@ -1,1 +1,1 @@
-github "OneSignal/OneSignal-iOS-SDK" "2.9.4"
+github "OneSignal/OneSignal-iOS-SDK" "2.14.3"
