+++ ios/Classes/ComWilliamrijksenOnesignalModule.h
@@ -13,6 +13,9 @@
 typedef void(^TagsResultHandler)(NSDictionary*, NSError*);
 
 - (void)promptForPushNotificationsWithUserResponse:(id)args;
+- (bool)retrieveSubscribed:(id)args;
+- (NSString *)retrievePlayerId:(id)args;
+- (NSString *)retrieveToken:(id)args;
 - (void)setSubscription:(id)args;
 - (void)setExternalUserId:(id)args;
 - (void)removeExternalUserId:(id)args;
