--- ios/platform/OneSignal.framework/Modules
@@ -1,1 +1,0 @@
-Versions/A/Modules
