--- ios/platform/OneSignal.framework/Resources
@@ -1,1 +1,0 @@
-Versions/A/Resources
